function moved to utils:
